+++ Example/Source/ViewController.h
@@ -1,7 +1,5 @@
 //
 //  ViewController.h
-//
-//  Copyright (c) 2013-2017 Evgeny Aleksandrov. License: MIT.
 
 #import <UIKit/UIKit.h>